Fix news articles not being orderable (#102)
This adds a rank column to News so that, if rank exists, it will use that to
prioritize postings.

This also fixes a bug affecting other models where the ordering was not as
expected. The scope name latest was not a good choice as latest() is an
existing Illuminate\Database\Eloquent\Model built-in that orderes the query by
created_at.