Add canonical image paths to most models (#169)
This commit adds ImageCache as a simple, first-pass way to handle
automatic resizing of images.

The URL for images associated with a model is:

    /{model}s/{id}/image/{preset}.{format}

Currently the cache is stored locally in storage/app/imagecache. In
the future, it will make more sense to store the cached images on
Digital Ocean Spaces so that it 1) can be served by their CDN
servers, 2) models can generate images after a new image upload
so we aren't doing generation per request, 3) HTML views can point
directly to the CDN URL and avoid having the PHP server handle the
image requests.

For RSS, we will want to keep serving images via the model URL so
that it can be stable.